requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.12
folium>=0.14
contextily>=1.4
jupyter>=1.0
notebook>=7.0
pyarrow>=14.0
shap>=0.43
pyyaml>=6.0
pytest>=7.0
